requirements.txt
anyio==4.0.0
APScheduler==3.6.3
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pyaes==1.6.1
pyasn1==0.5.0
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
python-telegram-bot==13.14
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
Telethon==1.30.0
tornado==6.1
tzlocal==5.0.1


